+++ packages/Core/lib/interfaces.js
@@ -1,5 +1,5 @@
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.VERSION = 2.11;
+exports.VERSION = 2.12;
 ;
 var InanoSQLFKActions;
 (function (InanoSQLFKActions) {