+++ packages/Core/CHANGELOG.md
@@ -4,6 +4,12 @@
 - Use [unfetch](https://github.com/developit/unfetch), [sockette](https://github.com/lukeed/sockette) and [Websocket Node](https://github.com/theturtle32/WebSocket-Node) for new client/server code.
 - Add SDL schema support. [Link](https://www.prisma.io/blog/graphql-sdl-schema-definition-language-6755bcb9ce51/).
 
+## [2.1.2] 1-30-2019
+- **BREAKING CHANGE** Removed `.streamEvent` function, you can now just pass a boolean into the `.stream` function to enable events: `.stream(onRow, complete, error, useEvents)`.
+- `objectsEqual` utility function is now an alias of the `fast-deep-equal` library, should have better performance for row comparisons and index updates.
+- Restored observer query feature as `.listen()` queries.
+- Adjusted RocksDB adapter so it can accept other leveldown adapters easily.
+
 ## [2.1.1] 1-24-2018
 - Added gitter chat link.
 - Added foreign key feature and its documentation.